requirements-dev.txt
flask==3.0.3
gunicorn==21.2.0
pyjwt==2.9.0
prometheus-client==0.20.0
requests==2.32.3
pytest==8.3.3
pytest-flask==1.3.0
flake8==7.1.1
black==24.10.0
